requirements.txt
Flask
SQLAlchemy
psycopg2-binary
marshmallow
python-dotenv
flask-cors
flask-migrate
flask-bcrypt
flask-restful
flask_jwt_extended
flask_mail
requests
Flask-SocketIO
python-engineio
python-socketio
eventlet
Flask-JWT-Extended
Flask-SQLAlchemy
cryptography
boto3
python-dotenv
schedule
pydrive2
cryptography
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
